アイテム一覧の取得実装完了

+++ ramen_shop_order_api/db/migrations/20200424160347_create_items.rb
@@ -4,7 +4,7 @@ Hanami::Model.migration do
       primary_key :id
 
       column :name, String, null: false
-      column :price, BigDecimal, null: false
+      column :price, Integer, null: false
 
       column :created_at, DateTime, null: false
       column :updated_at, DateTime, null: false
